Extract utils and constants
