+++ src/services/http/utils.ts
@@ -1,0 +1,12 @@
+import multer from 'multer';
+
+export const upload = multer({
+  dest: './uploaded_files',
+});
+
+export function makeResponse(payload: any, error?: boolean) {
+  return {
+    payload,
+    error,
+  };
+}
